requirements.txt
pytest
requests
pytest-html
google-cloud-bigquery
python-dotenv
